fix(compatibility): Fix for latest stable Steam client version

+++ src/feats/apps.hpp
@@ -18,7 +18,7 @@ namespace Apps
 	bool unlockApp(uint32_t appId, CAppOwnershipInfo* info);
 
 	bool checkAppOwnership(uint32_t appId, CAppOwnershipInfo* info);
-	void getSubscribedApps(uint32_t* appList, size_t size, uint32_t& count);
+	void getSubscribedApps(uint32_t* appList, uint32_t size, uint32_t& count);
 
 	bool shouldDisableCloud(uint32_t appId);
 	bool shouldDisableCDKey(uint32_t appId);
